+++ modules/self-managed-node-group/main.tf
@@ -460,8 +460,7 @@ resource "aws_security_group" "this" {
   tags = merge(
     var.tags,
     {
-      "Name"                                      = local.security_group_name
-      "kubernetes.io/cluster/${var.cluster_name}" = "owned"
+      "Name" = local.security_group_name
     },
     var.security_group_tags
   )