+++ outputs.tf
@@ -165,10 +165,10 @@ output "aws_auth_configmap_yaml" {
   description = "Formatted yaml output for base aws-auth configmap containing roles used in cluster node groups/fargate profiles"
   value = templatefile("${path.module}/templates/aws_auth_cm.tpl",
     {
-      eks_managed_role_arns        = [for group in module.eks_managed_node_group : group.iam_role_arn]
-      self_managed_role_arns       = [for group in module.self_managed_node_group : group.iam_role_arn if group.platform != "windows"]
-      win32_self_managed_role_arns = [for group in module.self_managed_node_group : group.iam_role_arn if group.platform == "windows"]
-      fargate_profile_arns         = [for group in module.fargate_profile : group.fargate_profile_arn]
+      eks_managed_role_arns                   = [for group in module.eks_managed_node_group : group.iam_role_arn]
+      self_managed_role_arns                  = [for group in module.self_managed_node_group : group.iam_role_arn if group.platform != "windows"]
+      win32_self_managed_role_arns            = [for group in module.self_managed_node_group : group.iam_role_arn if group.platform == "windows"]
+      fargate_profile_pod_execution_role_arns = [for group in module.fargate_profile : group.fargate_profile_pod_execution_role_arn]
     }
   )
 }